refactor: Update file paths for system instructions and chat history
