+++ instructions.js.sample
